finished roles module

+++ Project/User.rb
@@ -1,6 +1,9 @@
 require_relative 'modules\Autentication.rb'
+require_relative 'modules\Roles.rb'
 class User
   @@users = []
+  include Roles
+
   def initialize(username, password, rol)
     @user = {
       :username => username,
@@ -22,6 +25,7 @@ class User
   end
 
   def self.remove_user(current_user)
+    return puts "no tienes permiso" if Roles.is_admin?
     index = find_user(current_user[:username])
 
     @@users.delete_at(index)
@@ -47,17 +51,18 @@ end
 #puts ""
 #
 #user = {username: 'pacoporros', password: 'pacoelporros'}
-#User.modify_user(user, 'paco-porros', 'paco-el-porros', 'admin')
+#User.modify_user(user, 'paco-porros', 'paco-el-porros', 'normal')
 #puts "all users but one modified #{User.all_users} "
-#puts""
+##puts""
 #
-#user = {username: 'paco-porros', password: 'paco-el-porros'}
-#User.remove_user(user)
-#puts "#{User.all_users}"
 #users = User.all_users
 #Autentication.login(users,'paco-porros','paco-el-porros')
 #
 #puts "this is the current user: #{Autentication.current_user}"
+#
+#user = {username: 'paco-porros', password: 'paco-el-porros'}
+#User.remove_user(user)
+#puts "#{User.all_users}"
 #Autentication.logout
 #puts "this is the current user: #{Autentication.current_user}"
 #puts "is loged? #{Autentication.is_loged?}"
